Rescue GdsApi::HTTPClientError and return bad request
In our Sentry error log there are numerous errors raised when clients
have tried to put in dubious strings in a query string which have then
been forwarded to internal Rails APIs [1].

These dubious query strings can result in the downstream Rails APIs
returning a 400 bad request response. It seems appropriate to then
return this to user.

One thing to note is that this will catch all other 4xx responses (such
as 422 or 401) that are not configured with their own exception
handlers. This doesn't seem too big a semantic problem since they are
types of bad request so meet the semantics. Were this to become a
problem we would want to expand GdsApiAdapters to have a dedicated
HTTPBadRequest exception.

[1]: https://sentry.io/organizations/govuk/issues/881296490/?project=202221&query=&statsPeriod=14d