Rename and extend verification feature specs
Previusly the "manage_subscriptions_confirmation_email_spec" was
particularly ambiguous, as the name suggested a confirmation email
that the user had managed their subscriptions in the past.

This renames the existing feature specs related to the sending and
processing of a verification email, adding a missing spec for the
latter - for logging in to manage subscriptions.