Update login check email screen to match design
This switches to using '_html' locales in order to avoid inventing
a key for each paragraph. Note that the body style for the details
component already adds spacing between paragraphs, so we don't need
any manual '<br>' elements.

In order to keep the specs passing, this also switches to checking
just the title of the 'check email' page, since none of the content
is critical to the feature.